moved component s from app.jsx to separate files
